requirements.txt
flask
openai
pyngrok
python-dotenv